requirements.txt
apache-beam[gcp]
google-cloud-storage==2.6.0